update chess from NSI

+++ PycharmProjects/Working/chess/src/data/chess_images/pieces_dowloader.py
@@ -1,6 +1,7 @@
 import requests
 import os
 
+print("Downloading Pieces png")
 # Correspondance code → nom anglais
 pieces = {
     "wp": "white_pawn",
@@ -23,9 +24,14 @@ for code, name in pieces.items():
     url = f"{base_url}{code}.png"
     response = requests.get(url)
     if response.status_code == 200:
-        path = f"{code}.png"  # Enregistre directement dans le répertoire courant
-        with open(path, "wb") as f:
-            f.write(response.content)
-        print(f"Téléchargé et enregistré : {path}")
+        path = f"{os.path.dirname(os.path.abspath(__file__))}\{code}.png"  # Enregistre directement dans le répertoire courant
+        if not os.path.exists(path):
+            with open(path, "wb") as f:
+                f.write(response.content)
+            print(f"Téléchargé et enregistré : {path}")
+        else:
+            print(f"Already exist : {path[-34:]}")
     else:
         print(f"Erreur téléchargement : {url}")
+else:
+    print("Download complete\n")
